Implement binary protocol

+++ src/data.py
@@ -20,11 +20,11 @@ class Transaction:
     number: int
     from_username: username_t
     to_username: username_t
-    timestamp: float
+    timestamp: int
 
     @classmethod
     def from_intent(cls, ti: TransactionIntent, current_trn: int, now_fn=None) -> "Transaction":
-        now_fn = now_fn or (lambda: datetime.utcnow().timestamp())
+        now_fn = now_fn or (lambda: int(datetime.utcnow().timestamp()))
         return cls(current_trn + 1, ti.from_username, ti.to_username, now_fn())
